requirements.txt
ortools>=9.7
supabase>=2.3
python-dotenv>=1.0